feat(ai): LLM API 호출 시와 전체 분석 완료 시 로깅

+++ src/main/java/com/example/konnect_backend/domain/ai/dto/response/RunResultResponse.java
@@ -1,5 +1,5 @@
 package com.example.konnect_backend.domain.ai.dto.response;
 
-public record RunResultResponse(String response, Integer latencyMs, Long inputTokens,
-                                Long outputTokens) {
+public record RunResultResponse(String response, Integer latencyMs, Integer inputTokens,
+                                Integer outputTokens) {
 }
